% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.manag.surv.R
\name{data.manag.surv}
\alias{data.manag.surv}
\title{Management of survival data}
\usage{
data.manag.surv(formGroup, formSurv, data.long1, formSurv_CompRisk)
}
\arguments{
\item{formGroup}{A formula which indicates the group variable}

\item{formSurv}{A formula which indicates the variables used in the survival submodel}

\item{data.long1}{Database}

\item{formSurv_CompRisk}{A formula which indicates the variables used in the competing survival submodel}
}
\value{
A list with the following components :
\describe{
\item{\code{tmp}}{the final database for survival analysis}
\item{\code{Time}}{a vector of observed times}
\item{\code{event1}}{a vector of first event indicator}
\item{\code{nTime}}{length of Time vector}
\item{\code{Z}}{matrix of covariables of first survival submodel}
\item{\code{event2}}{a vector of second event indicator}
\item{\code{Z_CR}}{matrix of covariables of second survival submodel}
}
}
\description{
Management of survival data
}
