% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.time.R
\name{data.time}
\alias{data.time}
\title{Management of data for longitudinal submodel}
\usage{
data.time(data.id, Time, formFixed, formRandom, timeVar)
}
\arguments{
\item{data.id}{A dataframe}

\item{Time}{A vector of Time of events}

\item{formFixed}{A formula for the fixed effects of the longitudinal submodel}

\item{formRandom}{A formula for the random effects of the longitudinal submodel}

\item{timeVar}{The name of the column of time in data.id. This variable must appears in data.id}
}
\value{
A list with the following components
\describe{
\item{\code{Xtime}}{a matrix of fixed effects at each time of measure}
\item{\code{Utime}}{a matrix of random effects at each time of measure}
}
}
\description{
Management of data for longitudinal submodel
}
