% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexVarJM-package.R
\docType{package}
\name{FlexVarJM-package}
\alias{FlexVarJM}
\alias{FlexVarJM-package}
\title{FlexVarJM: Estimate Joint Models with Subject-Specific Variance}
\description{
In a location-scale model framework, this package allows to estimate mixed models with a subject-specific variance, lsmm, lsjm, competing, frequentiste marquardt.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LeonieCourcoul/FlexVarJM}
  \item Report bugs at \url{https://github.com/LeonieCourcoul/FlexVarJM/issues}
}

}
\author{
\strong{Maintainer}: Léonie Courcoul \email{leonie.courcoul@u-bordeaux.fr}

Authors:
\itemize{
  \item Antoine Barbieri \email{antoine.barbieri@u-bordeaux.fr}
  \item Hélène Jacqmin-Gadda \email{helene.jacqmin-gadda@u-bordeaux.fr}
}

}
\keyword{internal}
