% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussKronrod.R
\name{gaussKronrod}
\alias{gaussKronrod}
\title{Gauss-Kronrod nodes and weights}
\usage{
gaussKronrod(k = 15)
}
\arguments{
\item{k}{the number of points for Gauss-Kronrod approximation : choice between 7 and 15. 15 by default.}
}
\value{
A list with the following components :
\describe{
\item{\code{sk}}{A k-vector of nodes}
\item{\code{wk}}{A k-vector of weights}
}
}
\description{
Gauss-Kronrod nodes and weights
}
