% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Predictions for the goodness of fit, of the random effects, the current value for each individuals and the cumulative hazard function for both events}
\usage{
goodness_of_fit(FlexVarJM, graph = F, break.times = NULL)
}
\arguments{
\item{FlexVarJM}{an object of class FlexVarJM}
}
\value{

}
\description{
Predictions for the goodness of fit, of the random effects, the current value for each individuals and the cumulative hazard function for both events
}
\examples{

}
