% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.long.R
\name{initial.long}
\alias{initial.long}
\title{Initialisation of Longitudinal Submodel}
\usage{
initial.long(formFixed, formRandom, idVar, data.long1, ncX, nproc = nproc)
}
\arguments{
\item{formFixed}{A formula which indicates the fixed effects for the longitudinal submodel}

\item{formRandom}{A formula which indicates the random effects for the longitudinal submodel}

\item{idVar}{A character, indicates the name of the group variable}

\item{data.long1}{A dataframe with the longitudinal data}

\item{ncX}{An integer, the number of columns of matrix X, ie, the number of fixed effects}

\item{nproc}{An integer, the number of cores for parallel computation}
}
\value{
A list with the following components :
\describe{
\item{\code{long_model}}{the result of the hlme function}
\item{\code{priorMean.beta}}{the estimated parameters for fixed effects in the linear mixed effects model}
\item{\code{sigma}}{the estimated sigma of the model}
}
}
\description{
Initialisation of Longitudinal Submodel
}
