% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_s.t.bootstrap.R
\name{pred_s.t.bootstrap}
\alias{pred_s.t.bootstrap}
\title{Bootstrap for the predictions}
\usage{
pred_s.t.bootstrap(newdata, FlexVarJM, s, window, event = 1, L)
}
\arguments{
\item{newdata}{}

\item{FlexVarJM}{}

\item{s}{}

\item{window}{}

\item{event}{}
}
\description{
Bootstrap for the predictions
}
