% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_s.t.ponctuel.R
\name{pred_s.t.ponctuel}
\alias{pred_s.t.ponctuel}
\title{Predictions computation}
\usage{
pred_s.t.ponctuel(newdata, FlexVarJM, s, window, event = 1)
}
\arguments{
\item{newdata}{}

\item{FlexVarJM}{}

\item{s}{}

\item{window}{}

\item{event}{}
}
\description{
Predictions computation
}
