% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_s.t.R
\name{pred_s.t}
\alias{pred_s.t}
\title{Predictions computation}
\usage{
pred_s.t(object, newdata, s, window, event = 1, IC = NULL, draws = NULL)
}
\arguments{
\item{newdata}{}

\item{s}{}

\item{window}{}

\item{event}{}

\item{x}{}
}
\description{
Predictions computation
}
