% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.re.R
\name{predict.re}
\alias{predict.re}
\title{predict.re : a function to compute the predicted random effects of the database.}
\usage{
\method{predict}{re}(object)
}
\arguments{
\item{object}{an object of class 'lsjm'}
}
\description{
predict.re : a function to compute the predicted random effects of the database.
}
