% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_re.R
\name{predict_re}
\alias{predict_re}
\title{Function to predict the random effects}
\usage{
predict_re(
  param,
  nb.e.a = 0,
  X_base_i = NULL,
  beta = NULL,
  U_i = NULL,
  y_i = NULL,
  Sigma.b = NULL,
  mu.log.sigma = NULL,
  tau.log.sigma = NULL,
  variability_hetero = TRUE,
  alpha.sigma = NULL,
  competing_risk = FALSE,
  alpha.sigma.CR = NULL,
  sharedtype = NULL,
  sharedtype_CR = NULL,
  Xtime_i = NULL,
  Utime_i = NULL,
  Xs_i = NULL,
  Us_i = NULL,
  alpha.current = NULL,
  alpha.current.CR = NULL,
  Xslope_i = NULL,
  Uslope_i = NULL,
  Xs.slope_i = NULL,
  Us.slope_i = NULL,
  alpha.slope = NULL,
  alpha.slope.CR = NULL,
  indices_beta_slope = NULL,
  hazard_baseline = NULL,
  wk = NULL,
  shape = 0,
  Time_i = NULL,
  st_i = NULL,
  gamma = NULL,
  B_i = NULL,
  Bs_i = NULL,
  Z_i = NULL,
  alpha = NULL,
  P_i = NULL,
  hazard_baseline_CR = NULL,
  shape.CR = NULL,
  gamma.CR = NULL,
  B.CR_i = NULL,
  Bs.CR_i = NULL,
  Z.CR_i = NULL,
  alpha.CR = NULL,
  event1_i = NULL,
  event2_i = NULL
)
}
\arguments{
\item{event2_i}{}
}
\value{

}
\description{
Function to predict the random effects
}
\examples{

}
