% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_gf.R
\name{prediction_gf}
\alias{prediction_gf}
\title{Predictions for the goodness of fit, of the random effects, the current value for each individuals and the cumulative hazard function for both events}
\usage{
prediction_gf(FlexVarJM)
}
\arguments{
\item{FlexVarJM}{an object of class FlexVarJM}
}
\value{

}
\description{
Predictions for the goodness of fit, of the random effects, the current value for each individuals and the cumulative hazard function for both events
}
\examples{

}
