% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predyn.R
\name{predyn}
\alias{predyn}
\title{Dynamic prediction for a new individual}
\usage{
predyn(newdata, FlexVarJM, s, window = 1, event = 1, L = 500, graph = F)
}
\arguments{
\item{graph}{}
}
\description{
Dynamic prediction for a new individual
}
