% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexVarJM-package.R
\docType{package}
\name{FlexVarJM-package}
\alias{FlexVarJM}
\alias{FlexVarJM-package}
\title{FlexVarJM: Estimate Joint Models with Subject-Specific Variance}
\description{
Estimation of mixed models including a subject-specific variance which can be time and covariate dependent. In the joint model framework, the package handles left truncation and allows a flexible dependence structure between the competing events and the longitudinal marker. The estimation is performed under the frequentist framework, using the Marquardt-Levenberg algorithm. (Courcoul, Tzourio, Woodward, Barbieri, Jacqmin-Gadda (2023) \href{https://arxiv.org/abs/2306.16785}{arXiv:2306.16785}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LeonieCourcoul/FlexVarJM}
  \item Report bugs at \url{https://github.com/LeonieCourcoul/FlexVarJM/issues}
}

}
\author{
\strong{Maintainer}: Léonie Courcoul \email{leonie.courcoul@u-bordeaux.fr}

Authors:
\itemize{
  \item Antoine Barbieri \email{antoine.barbieri@u-bordeaux.fr}
  \item Hélène Jacqmin-Gadda \email{helene.jacqmin-gadda@u-bordeaux.fr}
}

}
\keyword{internal}
