% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarJM-package.R
\docType{package}
\name{FlexVarJM-package}
\alias{FlexVarJM}
\alias{FlexVarJM-package}
\title{FlexVarJM: Estimate Joint Models with subject-specific variance}
\description{
This package allows to estimate joint models with a subject-specific variance.
}
\author{
\strong{Maintainer}: Léonie Courcoul \email{leonie.courcoul@u-bordeaux.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
