\name{FlexVarJM-package}
\alias{FlexVarJM-package}
\alias{FlexVarJM}
\docType{package}
\title{
\packageTitle{FlexVarJM}
}
\description{
\packageDescription{FlexVarJM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FlexVarJM}
\packageIndices{FlexVarJM}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{FlexVarJM}

Maintainer: \packageMaintainer{FlexVarJM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
