% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexVarJM-package.R
\docType{package}
\name{FlexVarJM-package}
\alias{FlexVarJM}
\alias{FlexVarJM-package}
\title{FlexVarJM: Estimate Joint Models with subject-specific variance}
\description{
This package allows to estimate joint models with a subject-specific variance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LeonieCourcoul/FlexVarJM}
  \item Report bugs at \url{https://github.com/LeonieCourcoul/FlexVarJM/issues}
}

}
\author{
\strong{Maintainer}: Léonie Courcoul \email{leonie.courcoul@u-bordeaux.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
